+++ R/tidy_xlsx.R
@@ -83,6 +83,8 @@
 #' * `numeric` The numeric value of a cell.
 #' * `date` The date value of a cell.
 #' * `character` The string value of a cell.
+#' * `character_formatted` A data frame of substrings and their individual
+#'     formatting.
 #' * `formula` The formula in a cell (see 'Details').
 #' * `is_array` Whether or not the formula is an array formula.
 #' * `formula_ref` The address of a range of cells group to which an array
@@ -140,6 +142,12 @@
 #'   the RGB values in a spreadsheet program (e.g. Excel, LibreOffice,
 #'   Gnumeric), and use the [grDevices::rgb()] function to convert these to a
 #'   hexadecimal string.
+#'
+#'   Strings can be formatted within a cell, so that a single cell can contain
+#'   substrings with different formatting.  This in-cell formatting is available
+#'   in the column `character_formatted`, which is a list-column of data frames.
+#'   Each row of each data frame describes a substring and its formatting.
+#'   When a particular format is `NA`, the overall cell format applies.
 #' }
 #'
 #' @export
@@ -171,6 +179,11 @@
 #' # the relevant indices, and then filter the cells by those indices.
 #' bold_indices <- which(x$formats$local$font$bold)
 #' Sheet1[Sheet1$local_format_id %in% bold_indices, ]
+#'
+#' # In-cell formatting is available in the `character_formatted` column as a
+#' # data frame, one row per substring.  When a particular format is `NA`, the
+#' # overall cell format applies.
+#' tidy_xlsx(examples)$data$Sheet1$character_formatted[77]
 #' }
 tidy_xlsx <- function(path, sheets = NA) {
   .Deprecated(msg = paste("'tidy_xlsx()' is deprecated.",